% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_enrichment.R
\name{kegg_metabolite_enrichment}
\alias{kegg_metabolite_enrichment}
\title{Perform KEGG metabolite enrichment}
\usage{
kegg_metabolite_enrichment(KEGGid, species, p.adjust.method = "BH")
}
\arguments{
\item{KEGGid}{A character vector of KEGG compound IDs (e.g., "C00160")}

\item{species}{A character string, KEGG species code (e.g., "hsa", "eco")}

\item{p.adjust.method}{Adjustment method for multiple testing, default is "BH"}
}
\value{
A data.frame with enrichment results
}
\description{
Perform KEGG metabolite enrichment
}
