% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_plot.R
\name{keggbar}
\alias{keggbar}
\title{Title: Draw KEGG dotplot}
\usage{
keggbar(results)
}
\arguments{
\item{results}{A data.frame returned from kegg_metabolite_enrichment().}
}
\value{
A ggplot2 object.
}
\description{
Draw a barplot of KEGG enrichment results returned by `kegg_metabolite_enrichment()`.
}
